Plugin to display the open files upon test failure
The plugin is enabled by default. It displays the path to the files that
have a open file handler. To disable the plugin, run mda_nosetests with
the --no-open-files option or with the NOSE_NO_MDA_FILES environment
variable set.

The commit also introduces a test that fails if the MDA_FAILURE_TEST
environment variable is set. This test comes handy to test nose behavior
upon failure.

The open-files plugin requires psutil. That requirement has been added
to setup.py.