Plugin that list leaving Universes after tests
The plugin can be enable by running mda_nosetests with the
`--with-universe` option or with the NOSE_MDA_UNIVERSE environment
variable.

The plugin lists the leaving Universes, so as the reader they use and
the references that still exist to these universes.

The plugin is disabled by default because it is quite verbose, and
because it relies on the gc module which availability may depends on the
python implementation.