DATAWriter bond types and Masses made valid
Writer raises an exception if there are bonds but bond types are not
_explicitly defined_ as integers >= 1, i.e. will raise an exception
given the default output of `bond.type`.

The Masses section is written for all atomtypes up to the maximum
atomtype represented in the written selection. If the mass for an
atomtype cannot be determined from the selection's universe, then it
is taken to be 1.