Analysis progmeter (#892)
* PEP8 Changes

* Add progressmeter to AnalysisBase

This also adds an initialization function to AnalysisBase. This is done
now because we need to do more then just to setup the iteration.

* Adopt Contacts

* Follow best practices for matplotlib

Returning an axis object is considered to be better since the user is
able to adjust the plot by himself if he wishes to do so.

* Addopt Persistencelength

* Adopt InterRDF

* Adopt LinearDensity

* Adopt AlignTrj

* update changelog

* Ensure old API still works

Since user code might already exist we keep make sure the old API still
works as well.