adjusted HOLE open file descriptor test (#129) for Darwin
In order to provoke the failure described by issue #129 we artificially lower the open file descriptors.
On Mac OS X on travis this always leads to failing tests, as discussed under
https://github.com/MDAnalysis/mdanalysis/pull/901#issuecomment-231938093 ; this hack increases the allowed
open fds when we are running on Darwin. According to @jbarnoud, this problem is solved by #363 so once that
corresponding PR #815 is merged we should be able to revert this commit.