iQIcBAABCAAGBQJXj88NAAoJEObgNkBNX4V0xjkP/jJvd2bevjPTFHKOspDErTtm  NIMUstA8lOQd9w5kL+PcYC3o+J7vh1/gjIyJsZ5iE9y9zQ3aYtzFQShPnCfrKYyl  Bco/xr+U3bN9piXTZqhYKIRzVUs6SSd2uy7q63de8QDNsNWkKouKZ1+PhvZPkMNN  i+PhBW/gp+PXeta+4Y5REnBrUpX4bW3DCHuKTJ+nM80PXmsMG0i64ShRX/umXnoG  qpBfOGfnr40SD/ZFmmYc8qqZvllzPjew8GMGXXdjidetOZHaAkFRDMzOr3FNYkWD  2zSKN95CJGDynSwdsDTo9rrUN5lcJr58JG+JeDBLoK7XxN5VVhPge+cV0gHF2SLk  TQcgRXXY/AJY0ZpME0PRk7YKpUPqW/UjKPHENRFqNPlskF+8dzvFu2KNk928A+Ws  4otsRWCzn76tzXsGAK66kXsX9l2mc3IxhDy9TH69GXhog7ASHlglJZ1kZTcPPR8T  h7SfjW1Pfi7C/HL7RJ+shmmIQd4qvfn828JK/SoHud6dG7/wmfqYIiva7RB0usdt  CK8y7an2XUGXjnvIv2C2CDUHKy0KgtdRw6wpwChkeVl9ci59cT0vHkOZqXgGSaBS  L168OTkm0djbpjrMGj8vWe3lna1/Fxf+ELYOC/rUTPOYt7T0SuRoUHhcHhURCD/w  vViD4Ic6nvs7OR+ODkyZ  =eUcw  -----END PGP SIGNATURE-----
adjusted HOLE open file descriptor test (#129) for Darwin

In order to provoke the failure described by issue #129 we artificially lower the open file descriptors.
On Mac OS X on travis this always leads to failing tests, as discussed under
https://github.com/MDAnalysis/mdanalysis/pull/901#issuecomment-231938093 ; this hack increases the allowed
open fds when we are running on Darwin. According to @jbarnoud, this problem is solved by #363 so once that
corresponding PR #815 is merged we should be able to revert this commit.