Beginnings of marking old test_atomgroupy tests.
We want to preserve old AtomGroup tests for now to make it clear what is
breaking and what is not from old behavior. I've begun marking each test
with either VALID or INVALID. VALID means that the test should pass
under the new system. INVALID means that the test should not pass, and
so the test should be altered or removed.

Tests that fail in either case I am marking with `pytest.mark.skip`
so that we can continue moving forward with tests that are failing and
haven't been reviewed yet.