ProgressMeter accepts both new and legacy formats
A previous commit, translated the formats from the legacy to the new
syntax. This would break user's code.

This commit adds a keyword `format_handling`. When the keyword is set to
"new", the format string is handled with the new {}-based syntax. When
the keyword is set to "legacy", the format string is handled with the
old %-based syntax. By default, the keyword is set to "auto"; then the
kind of syntax is guessed.