Add ProgressMeter.start
The new ProgressMeter.start method starts the clock for the progress
meter. This allows to account for the first iteration when measuring the
elapsed time; therefore allowing a more accurate estimate time for
completion.

The ProgressMeter continues to work without a call to start. Then, the
first update of the progress meter is counted as the initial time. This
typically ignores the first iteration in the elapsed time.

A call to start is added in AnalysisBase.run.

Note that start could be called directly in ProgressMeter.__init__. Yet,
a progress meter can be instantiated long before it actually get used
(especially in a notebook), causing bogus time tracking.