More flexible format in ProgressMeter
The progress meter now has a new *progress* attribute that represent a
visual display of the progress as a progress bar. This bar can be added
to the format using '{progress}'. Detailed formating of the progress bar
can be set using the format spec part of the format string. Thus,
'{progress:-25}' will display a progress bar spanned over 25 characters;
'{progress:*--80}' will display the bar spanned over 80 character, using
'*' to display the achieved fraction of the bar, and '-' for the
non-achieved fraction.

The *elapsed* and *time_per_iteration* attributes are now instances of
the _TimeDelta custom class. This new class gives easier access to the
time in various units, nd allows better formatting. For instance, the
elapsed time expressed in seconds can be displayed using
'{elapsed.in_seconds}' or '{elapsed:%s}'. '{elasped:%h:%M:%S}' will
display the elapsed time as hours:minutes:seconds.