Deprecate quiet in AnalysisBase
AnalysisBase._quiet remains for compatibility. Its use does not trigger
a deprecation warning as it is not a documented API; it could issue a
warning if defined as a property but such definition would be heavier.

Instructions on how to deal with the post deprecation phase are added in
the docstring of the new private function lib.log._set_verbose. Thet
function handles most of the logic behing the deprecation of the quiet
keyword. It issues the deprecation warning if needed, raises an
exception if the user provided contradictory values of quiet and
verbose, and gives the value of verbose based on verbose and quiet.