Deprecate quiet in hbond_analysis
MDAnalysis.analysis.hbonds.hbond_analysis is a peculiar case for the
deprecation of the *quiet* keyword argument in favor of *verbose*.
Indeed, the *verbose* keyword argument is already in use to toggle the
debug outputs.

Here we deprecate the use of *quiet* in favor of *verbose*, and we
deprecate the use of *verbose* in favor of *debug*. A deprecation
warning is issued in two cases:

* the user creates an instance of HydrogenBondAnalysis using the
  *verbose* keyword argument. In this case, the user wants to set up the
  debug mode and should have used *debug* in the new version.
* the user calls HydrogenBondAnalysis.run with the *quiet* argument.
  There, the user wants to toggle the progress meter and should use
  *verbose*.

If the user calls HydrogenBondAnalysis.run with the *verbose* keyword
argument, then we cannot now if he intended to toggle the progress meter
following the new usage, or to set the debug output following the old
usage. Therefore, such call cannot issue a deprecation warning. Yet, the
change in behavior is visible as soon as the first step of the analysis.