lib.util.Namespace raises the expected exception
`lib.util.Namespace` allows to query a namespace by index or by
attribute. It allows to use both the following syntaxes to access the
same element:

    ns = lib.util.Namespace()
    ns['key'] = 'plop'

    ns['key']  # Access by index
    ns.key  # Access by attribute

The class is used to permit some flexibility when accessing auxiliary
data attached to trajectories.

Prior to this commit, accessing a non-existing item from a namespace by
attribute would raise a KeyError while an AttributeError should be
raise. This breaks functions that test the features of an object by
looking for some attributes, and are therefore capturing AttributeError
as a sign the attribute is missing.

In particular, raising the wrong exception confused
`numpy.lib.type_check.iscomplexobj` in numpy 1.12.0. This confusion lead
to `assert_equal` to miss behave in that version of numpy, therefore
breaking the tests for MDAnalysis when upgrading from python 1.11.3 to
1.12.0.

This commit makes sure that the expected exception is raised when
accessing an non-existing item by attribute in a Namespace object.