Allow increasing intervals in the progress meter
While it is usefull to have the progress meter updated every step at the
beginning of the analysis to make sure everything is right, it becomes
less and less crucial as the analysis progresses. Progressive intervals
are used in gromacs progress meters. This commit introduces the same
interval scheme in the ProgressMeter class. To activate it, set the
*interval* keyword to "progressive" instead of a number of steps.

In the "progressive" scheme, all steps are displayed until step 20, then
every 10 steps are displayed until step 200, every 100 steps until step
2000, and every 1000 steps afterward.