Fix transfer_to_memory progress meter call
Since I merged #1030, Universe.transfer_to_memory could instanciate a
ProgressMeter with an invalid interval. By default, that interval was
set to None, which is an invalid interval since interval must be an
integer (or must be castable to an int).

Here, I make sure Universe.transfer_to_memory always use 1 as interval
for the progress meter. This leads to the same behaviour as before #1030
was merged: the progress meter updates at every frame transferred. Note
that this behavior is not the default behavior of the progress meter,
and that the default behavior may be preferable.

Fixes #1030