Fixes #1227 (#1228)
At Universe creation, the coordinates were not read when the topology
was provided as a Topology object. Skipping the coordinates was likely
there to by accident because of the case where only the topology is
given. This commit moves the skip, so trying to read coordinates is
skipped only when no coordinates are provided: we do not try to read
coordinates from a Topology object.