refactor libdcd header API
We now don't expose the CHARMM bitfield directly anymore in the header. The
bitfield is a internal implementation detail of readdcd.h and the wrong value to
expose to a user of libdcd. Rather we expose `is_periodic` to enable
reading/writing box information. With this change the write/read header
test symmetric now.