Ignore `exec-used` pylint warning in a test
The exec built-in is used in the tests about the deprecation warnings
for the instant selectors. This causes a warning from pylint as `exec`
can be a security concern in some context. Here, the input given to
`exec` is fully controlled and exec is not more dangerous than typing
the input instructions directly. As `exec` improves the readability of
the tests, this commit disable the warning for the two specific use of
`exec`.