Fix deprecation warning missfires (#1477)
* Fix deprecation warning missfires

`u.segments` and `u.residues` should not issue a deprecation warning.
This commit applies the same strategy than
a9f9655a3b45b8486a1aadb5ad903f8bfe68ef3f to avoid the unwanted warning.

* Ignore `exec-used` pylint warning in a test

The exec built-in is used in the tests about the deprecation warnings
for the instant selectors. This causes a warning from pylint as `exec`
can be a security concern in some context. Here, the input given to
`exec` is fully controlled and exec is not more dangerous than typing
the input instructions directly. As `exec` improves the readability of
the tests, this commit disable the warning for the two specific use of
`exec`.