Add moltypes as topology attribute for TPR
The TPR parser now exposes the molecule type as the moltypes topology
attribute. The moltypes attribute is defined at the residue level, but
can be accessed at the atom level as well; it behave in the same way as
the resnames attribute.

The molecule type is the name of the molecule to which the atom/residue
belongs. Two molecules that share a molecule type has the same topology
template, so by definition they will have the same number of atoms,
residue, bonds, angles ... On the contrary to segments, moltypes do not
have to be contiguous so multiple occurrences of the same moltype can be
defined in a topology with other molecule types in between; this happens
for instance in lipid bilayer simulations where the lipids are sorted by
leaflet.

In the specific case of gromacs, the molecule type is defined as the
first field in the `[moleculetype]` section of an ITP file. Only the TPR
parser implements the moltype attribute, though the attribute is generic
in principle and can be implemented in any other parser assuming the
format can name molecules.