fix header length information in dcd
A DCD has two byte positions to store the length of the remark section. First a
it reads number of bytes in the remark plus 4. Afterwards it has the number of
80 character sections. These two numbers have to match. So far we wrote first
that we use 164 chars and three 80 character blocks, 164 != 80*3 + 4

Now we corrected that and state correctly we write 244 bytes in the remark
section. We never noticed that because our reader ignores the information about
the byte length on looks only at the number of blocks.

As a side note here. DCD remarks always have have to be a multiple of 80 in
length.