Fixed MPI fork() warning when importing MDAnalysis after MPI_Init() call
Solved problem:
---------------
When used in an MPI environment, importing MDAnalysis may trigger an MPI
warning because importing the uuid module in MDAnalysis.core.universe
triggers a call to os.fork(). This happens if MPI_Init() has been called
prior to importing MDAnalysis.

Why this was fixed:
-------------------
For the user, it is completely unclear where and why the MPI warning was
triggered and whether that is of any concern.

How to reproduce:
-----------------
With OpenMPI (version should be irrelevant, tested on 1.10.2) and Python < 3.7, executing
the command
  echo "import mpi4py.MPI, MDAnalysis" | mpirun --mca mpi_warn_on_fork 1 -n 1 python -
yields the following output:

"""
An MPI process has executed an operation involving a call to the
"fork()" system call to create a child process.  Open MPI is currently
operating in a condition that could result in memory corruption or
other system errors; your MPI job may hang, crash, or produce silent
data corruption.  The use of fork() (or system() or other calls that
create child processes) is strongly discouraged.  

The process that invoked fork was:

  Local host:          <hostname> (PID <pid>)
  MPI_COMM_WORLD rank: 0

If you are *absolutely sure* that your application will successfully
and correctly survive a call to fork(), you may disable this warning
by setting the mpi_warn_on_fork MCA parameter to 0.
"""

If you cannot reproduce this warning, it is probably due to your Python
installation using an implementation of fork() that is named differently and,
thus, is not recognized by mpirun (can be checked for example with gdb).

Explanation:
------------
The problem is actually caused by the uuid module, which is imported in
MDAnalysis.core.universe and triggers a call to os.fork() by calling a couple
of methods from the ctypes module which in turn use the subprocess module
(that's where os.fork() actually gets called).
This means that the real problem is not due to MDAnalysis itself, but it also
means that it will probably never be fixed for older Python versions.
In Python 3.7, the problem with the uuid module is solved. (see
https://bugs.python.org/issue11063 for further information)
Luckily, for Python < 3.7, the uuid module works perfectly fine with
os.fork() disabled during its import.
A clean fix is therefore simply to disable os.fork() prior to importing the
uuid module in MDAnalysis.core.universe and to re-enable it afterwards (for
Python < 3.7 only).