Add a faster function to deduplicate indices
`mda.lib.cutil.unique_int_1d` takes a 1D array of integers and returns
an array of the unique values. It has similar perormances as
`np.unique` for randomly ordered values, but is faster for sorted values.

MDAnalysis needs to deduplicate arrays of indices for basic operations.
In most cases the indices are sorted so the deduplication can be done
in linear complexity. However, `np.unique` is a general function; it
systematically sorts the values with a quicksort, which makes the
complexity in the order of n log n.

`mda.lib.cutil.unique_int_1d` only works on 1D arrays of integer and
does not have the optional return values `np.unique` has. It assumes
first that the values are sorted and only sorts them if they happen not
to be.