Changed expected output from 0.0 to Nan.
The survival probability is a measure of how likely a selected molecule
is to remain within a specified region over a period of time. If at t0
the selected molecule is not in the specified region, no survival
probability can be calculated.

For example, consider that we would like to calculate the survival
probability of an ion at an interface. If the ion never approaches the
interface, the previous test would return a survival probability of 0,
which suggests that the ion comes to the interface but leaves
immediately. However, as the ion was never at the interface, no
survival probability can be calculated from this data.