Fix SETTLE in TPR parser
Support for SETTLE interactions in the TPR parser was added in #1953.
Though, the support was only added for the older versions of the TPR
format, causing errors with the newest files. In practice, when reading
the bonds on a recent TPR file containing SETTLE constraints, the
indices to the atoms involved in the bonds would be shifted. This is
because, the older TPR files use a different format to describe SETTLE
constraints.

This commit should fix the issue by supporting both the old and the new
format for SETTLE.