Extracting the autocorrelation and intermittency into their own functions with their own test cases.
Removed the class HydrogenBondLifetimes and moved the autocorrelation
functionality into the hbond_analysis class as a .autocorrelation
method. This method should replace also hbond_autocorrel.py file. Test
cases for the new method .autocorrelation are added.

Added a test cases that, with our definition
of intermittency, would fail in the case of the hbond_autocorrel.py.