Introduce a new unpacker for TPR files
Starting with gromacs 2020, the body of TPR files is not encoded in XDR
format. While the API to read the file remains the same, the underlying
encoding is different: instead of using file-byte per data-byte, the new
format uses 4 file-bytes per data-byte, the new format is also
little-endian rather than big-endian like before.