Providing Elements attribute to PDB parser  (Issues #2553 & #2467) (#2627)
Fixes #2647, towards #2553 

## Work done in this PR
Elements are now read from PDB files and added to the elements attribute of the universe.
This is currently only done in cases where all elements in the PDB file are valid (i.e. match IUPAC naming). In all other cases a user warning is thrown and the elements are not assigned.

## Files changed
  - MDAnalysis/topology/PDBParser.py
  - MDAnalysisTests/topology/test_pdb.py
  - CHANGELOG and AUTHORS (new contributor)