Add elements attribute to TPR parser
Elements are read from the TPR. TPR files contain the atomic number for
each atoms, when an atom does not match an element (some virtual sites,
coarse-grained particles), the atomic number is set to -1.

The TPR parser adds the 'elements' attribute only if at least 1 atom has
its atomic number correspomding to an element. In that case, atoms that
do not match an element have theire element set to an empty string.

See #2553
Closes #2628