Use a ficture for npz files in test_bat
Fixes #2979

The file test_bat_IO.npy was created in one test and used in
another one. This leads to the second test possibly failing
when the tests are run in parallel. This commit moves creating
the file into a fixture that the two tests depends.

The new fixture also uses the tmpdir fixture to avoid the file
created during the tests to remain in the test directory.