Remove useless branch in PDB.py
There was a test in PDBWriter._write_pdb_bonds about writing the bonds
for an atomgroup that does not have AG.ids but asking to not reindex.
The test was silencing the problem and would lead to the bonds not being
written. However, that part of the code would never be reached because
writing the atoms would have raised an exception before that method is
called.

This commit removes this test. Now, if that part of the code is called
(which should not happen), an exception is raised when the method tries
to access AG.ids.