Coverage of different PBCs for distance benchmarks
Added, for relevant functions, benchmarks for testing orthogonal PBC
settings and also removed 10000 atom benchmarks for contact matrix
functions to avoid timeout failiures.

Had to split the DistancesBench class into 3 classes to avoid
failiures and uncessary reruns of the benchmark. These failiures
were due to time outs when too many atoms were run in the
benchmarks or distances.betweeen not taking box type as an argument.
The latter case meant, to avoid failiure, we had to either split
the classes out or run redundant benchmarks.

The 3 classes and the distribution of testing functions is

BetweenBench
    time_between
DistancesBench
    time_distance_array
    time_distance_array_pre_allocated
    time_self_distance_array
    time_self_distance_array_pre_allocated
    time_dist
    time_dist_offsets
ContactsBench
    time_contact_matrix
    time_contact_matrix_sparse

Splitting the classes like this was necessary because asv does not
allow us to easily skip certain combinations of parameters at the
function level, only in the setup()
function of the class.