Added `norm` attribute to rdf classes
This PR adds a `norm` attribute to the two rdf classes
`InterRDF` and `InterRDF_s`. It works similar as the density
parameter which was only available for `InterRDF_s`.

Also, the universe attribute in `InterRDF_s` is
depricated because the universe can be accessed via
the AtomGroups.

Finally, I slightly cleaned up the code and moved the
attributes within the classes themselves.