Reduce warnings from Universe.empty
Universe.empty raises a warning when atom_resindex or residue_segindex
are not provided. The warnings tell the user that the atoms will all be
assigned to the first residue or that the residues will all be assigned
to the first segment. These warnings are expecilly important so that
users are aware some residues or segments aill be empty.

However, if n_residues is 1 or n_segments is 1, the behaviour from the
warning is the intended one. Therefore, there should not be a warning.
This commit removes the warnings in these specific cases.