ENH: faster TPR reads
* hoist some unncessarily repeated operations in the TPR
topology construction loops

* `impropers`, `dihedrals`, `angles`, and `bonds` data
structures in the TPR construction now partially leverage
NumPy data structures and operations--danger--the testsuite
only appears to be sensitive to `bonds` and I intentionally
did not carry over the arithmetic from `remap_` functions if they
didn't cause problems in the testsuite--however, it seems likely
that we may have some testing issues to address before we can
be confident in perf enhancements here? Looks like more tests
for proper atom indices in those other data structures might
be needed? Or perhaps they get handled as a backup somewhere
else I suppose...

* I also didn't really pay much attention to anything else
the testsuite wasn't sensitive to--if a comment says that a topology
data structure is a bunch of tuples in a list, I felt free to
ignore that and use a NumPy data structure if the testsuite couldn't
detect a change

* large TPR local read-in time drops from 71 seconds to 65