ENH: support tpx 133 (GMX 2024.1)
* One of my postdocs (Emvia) needs parsing of newer
`.tpr` files to work; we actually are using tpx version
`130` internally, but that wasn't from a stable release,
so I just went for `2024.1` and presumbly we can just use
`convert-tpr` to "up-version" to the stable release version.

* This patch used `convert-tpr` with GMX `2024.1` to mimic
Jonathan's changes from gh-4052, when we last bumped `tpx`
version. The full testsuite passed for me locally at least, so
maybe we'll be lucky enough to have another simple bump.

* Note that I did not update the documentation or the `CHANGELOG`,
fell free to push that in here if you're confident of the various
tpx generation things, etc.