BLD: NumPy 2 compat for wheel builds (#4620)
* BLD: NumPy 2 compat for wheel builds

* Fixes #4619

* We should build against NumPy `2.0.0` so that we're backwards
compatible to NumPy `1.x` while still supporting 2.x series. This
is simpler to maintain as well. In theory, should support
as far back as NumPy `1.19` if we needed it.

* MAINT, TST: PR 4620 revisions

* Add NumPy 2-related shims aimed at avoiding
control flow that hits `RDKit` or `parmed`, until
they have releases supporting NumPy 2. This mostly
means skipping a bunch of tests that use those
packages for now.

* Update ParmEd.py

* Update CHANGELOG

* Update ParmEd.py

---------

Co-authored-by: Irfan Alibay <IAlibay@users.noreply.github.com>