MAINT: rounding residue partial charges
* Fixes gh-5032.

* Given that 1e-4 is already optimistic for electron charge unit
precision in MD forcefields, this patch proposes to round residue
net charges to 4 decimal places to avoid the confusion described
in the matching ticket.

* Some obvious questions might include: why are we only doing this
for AMBER topologies? why only round at the residue level and
not i.e., upstream at atom level and also at "segment" levels? This
is mostly for scoping, because this initial patch is targeted
at the specific issue raised, and because it probably doesn't make
sense to proceed much farther until things are discussed.