TST: PR 4873 revisions
* Adjust the testing in above PR such that we also probe coordinate
and velocity reading from TPR files using the "double incantation:"
`u = mda.Universe(tpr_file, tpr_file)`, and not just the single
incantation.

* Refer back to the above PR in the code comments for clarity,
based on reviewer feedback.

* Add a regression test for the raising of an appropriate error
when attempting to read positions and velocities from the beta
series of GROMACS 2020. There is part of the effort to improve
test coverage in the above PR.

* Add a simple regression test for the case of reading the topology
of a given system from a different version of TPR file than the
version of TPR containing the positions and velocities.

* Remove a guard for the presence of topology data in TPR
files--I'm not aware of any cases where the topology information
could reasonably be absent (no way to test this at the moment).

* Add full canonical file header to new files added in above PR.

* Update `CHANGELOG` to reflect the new functionality added in above PR.

* Add some initial/crude documentation to `TPR.py`, along with the
requested `versionadded` directive.