BUG: large DCD file seeking on Win
* Fixes gh-4879

* 32-bit signed integer arithemetic was being used
in the DCD trajectory seeking Cython logic on Windows,
causing an overflow for trajectories with a large number
of frames on that platform. The behavior has existed since
at least MDAnalysis `2.5.0`, and a regression test and
patch have been added here. The regression test is
based on the original scripts Oli provided in the matching
ticket, though I note in a comment that we likely don't actually
need a large file, and could probably simplify the regression
test by artificially setting the frame numbers to start from
a large value. Nonetheless, this test does fail before and
pass after the patch locally on Windows.